requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.2
PyPDF2==1.26.0
langchain==0.0.2  
nltk==3.6.5
pymongo==4.7.2
openai==1.30.1
distro==1.9.0
tiktoken==0.7.0
faiss-cpu==1.8.0